+++ src/components/cards/Widget/Capacity/components/CapacityRow.tsx
