+++ src/components/cards/styled.tsx
